refactored redundant code
