+++ gpt_researcher/master/agent/researcher.py
@@ -3,7 +3,8 @@ import random
 from typing import Dict, Optional
 
 from ..actions.utils import stream_output
-from ..actions import get_sub_queries, scrape_urls
+from ..actions import scrape_urls
+from ..actions.query_processing import get_sub_queries
 from ...document import DocumentLoader, LangChainDocumentLoader
 from ...utils.enum import ReportSource, ReportType, Tone
 
@@ -110,7 +111,7 @@ class ResearchConductor:
         """
         context = []
         # Generate Sub-Queries including original query
-        sub_queries = await self.__get_sub_queries(query)
+        sub_queries = await self.researcher.context_manager.get_sub_queries(query)
         # If this is not part of a sub researcher, add original query to research for better results
         if self.researcher.report_type != "subtopic_report":
             sub_queries.append(query)
@@ -142,7 +143,7 @@ class ResearchConductor:
         """
         context = []
         # Generate Sub-Queries including original query
-        sub_queries = await self.__get_sub_queries(query)
+        sub_queries = await self.researcher.context_manager.get_sub_queries(query)
         # If this is not part of a sub researcher, add original query to research for better results
         if self.researcher.report_type != "subtopic_report":
             sub_queries.append(query)
@@ -305,22 +306,3 @@ class ResearchConductor:
             self.researcher.vector_store.load(scraped_content_results)
 
         return scraped_content_results
-
-    async def __get_sub_queries(self, query):
-        await stream_output(
-            "logs",
-            "planning_research",
-            f"🌐 Browsing the web and planning research for query: {query}...",
-            self.researcher.websocket,
-        )
-
-        # Generate Sub-Queries including original query
-        return await get_sub_queries(
-            query=query,
-            retriever=self.researcher.retrievers[0],
-            agent_role_prompt=self.researcher.role,
-            cfg=self.researcher.cfg,
-            parent_query=self.researcher.parent_query,
-            report_type=self.researcher.report_type,
-            cost_callback=self.researcher.add_costs,
-        )
